# Run optimization algorithms over multiple cores