--- a/extern/fftw-3.3.4/fftw3.pc
+++ b/extern/fftw-3.3.4/fftw3.pc
@@ -1,0 +1,11 @@
+prefix=/export/sherpa/build
+exec_prefix=${prefix}
+libdir=${exec_prefix}/lib
+includedir=${prefix}/include
+
+Name: FFTW
+Description: fast Fourier transform library
+Version: 3.3.4
+Libs: -L${libdir} -lfftw3 
+Libs.private: -lm
+Cflags: -I${includedir}