--- a/extern/fftw-3.3.4/stamp-h1
+++ b/extern/fftw-3.3.4/stamp-h1
@@ -1,0 +1,1 @@
+timestamp for config.h