… .travis.yml
… extern/fftw-3.3.4/api/fftw3.f03
… extern/fftw-3.3.4/libbench2/libbench2.a
… extern/fftw-3.3.4/mpi/fftw3-mpi.f03
… extern/fftw-3.3.4/mpi/fftw3l-mpi.f03
… extern/fftw-3.3.4/tests/bench
… extern/fftw-3.3.4/tools/fftw-wisdom
… extern/fftw-3.3.4/tools/fftw-wisdom-to-conf
… extern/region-4.9/src/reglexer.c
… extern/region-4.9/src/regparser.c
… sherpa/astro/tests/test_astro.py
… sherpa/astro/ui/tests/test_serialize.py
… sherpa/optmethods/opt.py
… sherpa/optmethods/optfcts.py
… sherpa/optmethods/src/NelderMead.cc
… sherpa/optmethods/src/_saoopt.cc
… sherpa/optmethods/src/minim.cc
… sherpa/optmethods/src/minpack/LevMar.cc
… sherpa/optmethods/src/minpack/LevMar.hh
… sherpa/optmethods/tests/test_ncoresopt.py
… sherpa/optmethods/tests/test_optmethods.py
… sherpa/optmethods/tests/tstopt.hh
… sherpa/optmethods/tests/tstoptfct.hh
… sherpa/ui/tests/test_plot_pvalue.py
… sherpa/ui/utils.py