@olaurino 2019-02-13
@dtnguyen2, the PR is missing a Release Note.

It looks like a bunch of unintended/unrelated changes (and entire files) were included in the commits by mistake.

Also, a bunch of tests in this PR fail, a couple of which fail because a test file was removed, most likely by mistake.