@dtnguyen2 2019-03-22
See 2c in following doc:

https://docs.google.com/document/d/1wQmHUEviv2K4U5rSK3EH1hYO0ByAVY3SpbgkHG7T7RU/edit#