@anetasie 2019-06-12
I just run a simple image fitting test with `moncar` using a relatively large image (253,253) with no psf convolution. The execution time is 49 sec vs 100 sec when I use 3 cores instead of 1. I can definitely see the speed improved and get the same best fit values. This would be a good improvements for our image fitting.