@anetasie 2019-06-12
I noticed that the printout on the screen shows the number of function evaluation which is much smaller when I used 3 cores. I suspect that the printout to the screen shows the number of evaluations for one core, instead of 3, but the division of evaluations (3*3218=9654) does not match perfectly.

```
Dataset               = 1
Method                = moncar
Statistic             = cash
Initial fit statistic = -290594
Final fit statistic   = -515548 at function evaluation 8726
Data points           = 63756
Degrees of freedom    = 63751
Change in statistic   = 224954
   c2.c0          0.0680481   
   g1.fwhm        7.97214     
   g1.xpos        127.387     
   g1.ypos        127.061     
   g1.ampl        807.438     



Dataset               = 1
Method                = moncar
Statistic             = cash
Initial fit statistic = -290594
Final fit statistic   = -515548 at function evaluation 3218
Data points           = 63756
Degrees of freedom    = 63751
Change in statistic   = 224954
   c2.c0          0.0680001   
   g1.fwhm        7.97257     
   g1.xpos        127.387     
   g1.ypos        127.062     
   g1.ampl        807.188    

```