@olaurino 2019-06-27
Rebasing and force-pushing to resolve a (minor) conflict and get the correct sherpa-test-data commit.