@olaurino 2019-06-27
So all our tests are failing because anaconda changed the structure of their repositories (see https://www.anaconda.com/how-we-made-conda-faster-4-7/). I tested a solution locally but that will require rebasing a whole lot of PRs (or the PRs could cherry-pick the relevant commit, either way it's a lot of work).