@olaurino 2019-06-28
Rebased once more to get #636, hoping CI will finally run.