@olaurino 2019-06-28
Because of several issues, starting with the fact that the submodule's commits hadn't been pushed, the CI tests on this PR never executed. Now it looks like they are failing for Python 2.7:
https://travis-ci.org/sherpa/sherpa/jobs/551858697

I can reproduce the error on my workstation with Python 2.7 and this PR.