@olaurino 2019-06-28
Oh, and while the other tests haven't completed yet it looks like there are some other failures with Python 3 as well.