@olaurino 2019-07-12
There are still three errors in Travis:
  1. It looks like matplotlib 3 is issuing a new warning. I am inclined to open a new ticket to track it, but to whitelist it for the time being (https://travis-ci.org/sherpa/sherpa/jobs/557964942#L5326)
  2. Some tests for this PR are erroring out on macOS. Given the nature of the tests I think I'd skip them on macOS. As long as they run on Linux and pass, we are fine. The tests depend on the number of CPUs on the virtual machine (https://travis-ci.org/sherpa/sherpa/jobs/557964938#L3820)
  3. 1 test is failing with xspec 12.9.1. The test passes in all the other jobs. I am not sure what to make of that. A few days ago we almost ditched the 12.9.1 job. If we had, we wouldn't had seen this failure. I'd be inclined to skip the test (https://travis-ci.org/sherpa/sherpa/jobs/557964940)

@DougBurke thoughts? Especially on 3.