@DougBurke 2019-07-12
So, running things with some-level of multi-core code could explain XSPEC differences (if there are memory-read/write issues in the 12.9.1 code base, say). So I think we should just drop the 12.9.1 XSPEC tests.

For point 2, I am generally a bit wary of "oh, it's okay on linux, must be fine on macOS" style arguments, but I don't have the time to look into this particular case to see whether I'm being too picky.

For the matplotlib warning: I'd really prefer it if our "skip-this-warning" check were more targeted - i.e. we would mark up the particular tests with the warnings it is known to generate. I am just concerned that our current scheme means that we don't notice if a piece of code suddenly starts generating warnings that we've seen before. However, that is a larger issue than this PR.

I note we are still seeing the occasional

```
  /home/travis/miniconda/envs/build/lib/python3.7/importlib/_bootstrap.py:219: RuntimeWarning: Unable to load the ciao_version module to determine version number- defaulting 'group' version to 0.0.0
    return f(*args, **kwds)
```

warning.