@olaurino 2019-07-12
> For point 2, I am generally a bit wary of "oh, it's okay on linux, must be fine on macOS" style arguments, but I don't have the time to look into this particular case to see whether I'm being too picky.

This is not the case here. Note that jobs on Travis won't generally exercise parallelization because they run on docker containers and virtual machine that usually have only one core. Unless I am missing something (@dtnguyen2 please read this) the tests are just checking that the multicore code is running, the actual results are being tested in other parts of the test suite, and we did have some failures we had to address.

> For the matplotlib warning: I'd really prefer it if our "skip-this-warning" check were more targeted - i.e. we would mark up the particular tests with the warnings it is known to generate. I am just concerned that our current scheme means that we don't notice if a piece of code suddenly starts generating warnings that we've seen before. However, that is a larger issue than this PR.

Agreed. I wish we had more time for this kind of improvements, but all in all I think the current infrastructure is doing a good job at telling us when new warnings appear, as in this case. This is something we will have to fix before matplotlib 3.3 otherwise the warning will turn into an error.