@dtnguyen2 2019-07-12
When I originally checked the code in, @olaurino made a comment to me that if ```numcores != 1``` then how do you know that optimization routine was run on multi-cores?  So I added a check to verify that multi-cores were used.  The test for the fit results on different platforms are still being checked and they pass for all platforms. 

As for point# 2: I made the following changes to the code:
```
        if _ncpus != 1:
            assert fit_results.extra_output['num_parallel_map'] > 0
        else:
            assert fit_results.extra_output['num_parallel_map'] == 0
  ```
point #2 should no longer be an issue